feat(enmeshed): deploy own mongodb instead of builtin dev db
Somehow the container ferretdb, provided by the enmeshed helm chart,
was refusing connection to the connector container. Sometimes it healed
by itself, but then we had caching problem (since the db was also
deleted). Sometimes it didn't even healed by itself, so that we needed
to delete the pod.

Therefore, tt is more secure to have an extra mongodb instance.